+++ pkg/linters/const_group_switch_cover/const_group_switch_cover.go
@@ -12,21 +12,6 @@ import (
 	"github.com/ysuzuki19/robustruct/pkg/linters/robustruct/settings"
 )
 
-func TypeEqual(a, b types.Type) bool {
-	if a == nil || b == nil {
-		return false
-	}
-	if a == b {
-		return true
-	}
-	if namedA, ok := a.(*types.Named); ok {
-		if namedB, ok := b.(*types.Named); ok {
-			return namedA.Obj().Pkg() == namedB.Obj().Pkg() && namedA.Obj().Name() == namedB.Obj().Name()
-		}
-	}
-	return false
-}
-
 var Analyzer = &analysis.Analyzer{
 	Name: settings.FeatureConstGroupSwitchCover.String(),
 	Doc:  "checks that all cases in a switch statement with a constant group expression are full arms",
@@ -39,16 +24,6 @@ var Analyzer = &analysis.Analyzer{
 	Requires:         []*analysis.Analyzer{inspect.Analyzer},
 	ResultType:       nil,
 	FactTypes:        []analysis.Fact{},
-}
-
-func findImport(pass *analysis.Pass, name string) *types.Package {
-	for _, imported := range pass.Pkg.Imports() {
-		importedName := imported.Name()
-		if importedName == name {
-			return imported
-		}
-	}
-	return nil
 }
 
 func run(pass *analysis.Pass) (interface{}, error) {
